+++ connector/src/test/scala/quasar/fs/ManageFileSpec.scala
@@ -18,8 +18,8 @@ package quasar.fs
 
 import quasar.Predef._
 import quasar.Data
-import quasar.fp._, eitherT._
-import quasar.fp.ski._
+import quasar.contrib.scalaz.eitherT._
+import quasar.fp._, ski._
 
 import scala.Predef.$conforms
