+++ core/src/main/scala/quasar/fs/mount/hierarchical.scala
@@ -19,8 +19,8 @@ package quasar.fs.mount
 import quasar.Predef._
 import quasar.common.PhaseResults
 import quasar.contrib.pathy._
+import quasar.contrib.scalaz.eitherT._
 import quasar.effect._
-import quasar.fp._, eitherT._
 import quasar.fp.free._
 import quasar.fs._
 import quasar.frontend.{logicalplan => lp}, lp.LogicalPlan
