+++ couchbase/src/main/scala/quasar/physical/couchbase/fs/queryfile.scala
@@ -21,9 +21,10 @@ import quasar.{Data, DataCodec, RenderTreeT}
 import quasar.common.{PhaseResults, PhaseResultT}
 import quasar.common.PhaseResult.{detail, tree}
 import quasar.contrib.pathy._
+import quasar.contrib.scalaz.eitherT._
 import quasar.effect.{KeyValueStore, Read, MonotonicSeq}
 import quasar.effect.uuid.GenUUID
-import quasar.fp._, eitherT._
+import quasar.fp._
 import quasar.fp.free._
 import quasar.fp.ski._
 import quasar.fs._
