+++ effect/src/main/scala/quasar/errors.scala
@@ -17,7 +17,8 @@
 package quasar
 
 import quasar.Predef._
-import quasar.fp._, eitherT._
+import quasar.contrib.scalaz.eitherT._
+import quasar.fp._
 
 import scalaz._
 import scalaz.concurrent._
