+++ foundation/src/main/scala/quasar/fp/free/Interpreter.scala
@@ -17,7 +17,8 @@
 package quasar.fp.free
 
 import quasar.Predef._
-import quasar.fp._, eitherT._
+import quasar.contrib.scalaz.eitherT._
+import quasar.fp._
 
 import scalaz._
 import scalaz.stream.Process
