+++ it/src/test/scala/quasar/fs/ReadFilesSpec.scala
@@ -19,7 +19,8 @@ package quasar.fs
 import quasar.Predef._
 import quasar.Data
 import quasar.contrib.pathy._
-import quasar.fp._, eitherT._
+import quasar.contrib.scalaz.eitherT._
+import quasar.fp._
 import quasar.fp.numeric._
 
 import java.lang.RuntimeException
