+++ it/src/test/scala/quasar/physical/marklogic/qscript/XmlStructuralPlannerSpec.scala
@@ -18,7 +18,7 @@ package quasar.physical.marklogic.qscript
 
 import quasar.Predef._
 import quasar.Data
-import quasar.fp.eitherT._
+import quasar.contrib.scalaz.eitherT._
 import quasar.physical.marklogic.DocType
 import quasar.physical.marklogic.xml.QName
 import quasar.physical.marklogic.xquery._
