+++ it/src/test/scala/quasar/regression/QueryRegressionTest.scala
@@ -21,10 +21,11 @@ import quasar._
 import quasar.build.BuildInfo
 import quasar.common._
 import quasar.contrib.argonaut._
+import quasar.contrib.scalaz.eitherT._
 import quasar.ejson
 import quasar.frontend._
 import quasar.contrib.pathy._
-import quasar.fp._, eitherT._, free._
+import quasar.fp._, free._
 import quasar.fp.ski._
 import quasar.fs._
 import quasar.main.FilesystemQueries
